requirements.txt

fastapi>=0.111.0
uvicorn>=0.30.0
pydantic>=2.7.0
python-multipart>=0.0.9
pandas>=2.2.2
openpyxl>=3.1.2
pypdf>=5.0.0
python-docx>=1.1.0
chromadb>=0.5.3
openai>=1.37.0
